Include dependencies of Qt platform plugins
This should fix kasenvr/project-athena#516 and the error of:

    Could not load the Qt platform plugin "xcb" in "" even though it was found.

The issue happens due to a missing libxcb-xinerama.so.0, which isn't included
because plugins aren't direct dependencies and can't be found with ldd. Instead
we need to scan any dynamically loaded plugins even if they don't seem to be
referenced by anything.

This makes the resulting AppImage considerably larger.